+++ src/App.vue
@@ -1,8 +1,21 @@
 <template>
-  <div id="app">
-    <router-view />
+  <div id="app" class="flex flex-col h-full">
+    <div class="flex-grow">
+      <router-view />
+    </div>
+    <PageFooter />
   </div>
 </template>
+
+<script>
+import PageFooter from '@/components/ui/molecules/PageFooter'
+
+export default {
+  components: {
+    PageFooter,
+  },
+}
+</script>
 
 <style lang="sass">
 #app